QCamera2: Port HAL for 64 bit platforms
- Added -Wextra compiler warning level.
- Avoiding implicit datatype conversions
  and also pointer to integer conversions,
  including literals.
- Avoiding pointer conversions between different types
  other than (void*).
- Using 'size_t' to store memory size and indexes.
- Removed some unused variables.
- Changed datatype of some iterators to match their
  corresponding limiter types.
- Using double instead of float where possible.
- Changed bitmasks to unsigned types.
- Updated printf/scanf format string specifiers to match
  actual type of arguments.
- Using 'String8' instead of 'char[]' where possible
  at QCameraParameters.
- Checking result of QCameraMemory::getSize().
- Fixes build errors and warnings.

CRs-Fixed: 641033

+++ QCamera2/HAL/QCameraChannel.h
@@ -66,8 +66,8 @@ public:
     int32_t config();
     QCameraStream *getStreamByHandle(uint32_t streamHandle);
     uint32_t getMyHandle() const {return m_handle;};
-    uint8_t getNumOfStreams() const {return m_numStreams;};
-    QCameraStream *getStreamByIndex(uint8_t index);
+    uint32_t getNumOfStreams() const {return m_numStreams;};
+    QCameraStream *getStreamByIndex(uint32_t index);
     QCameraStream *getStreamByServerID(uint32_t serverID);
     int32_t UpdateStreamBasedParameters(QCameraParameters &param);
     void deleteChannel();
@@ -79,7 +79,7 @@ protected:
     bool m_bAllowDynBufAlloc; // if buf allocation can be in two steps
 
     uint32_t m_handle;
-    uint8_t m_numStreams;
+    uint32_t m_numStreams;
     QCameraStream *mStreams[MAX_STREAM_NUM_IN_BUNDLE];
     mm_camera_buf_notify_t mDataCB;
     void *mUserData;
@@ -121,7 +121,7 @@ public:
                                        cam_pp_feature_config_t &config,
                                        QCameraChannel *pSrcChannel,
                                        uint8_t minStreamBufNum,
-                                       uint32_t burstNum,
+                                       uint8_t burstNum,
                                        cam_padding_info_t *paddingInfo,
                                        QCameraParameters &param,
                                        bool contStream,
@@ -131,7 +131,7 @@ public:
             QCameraParameters &param);
 
     // offline reprocess
-    int32_t doReprocess(int buf_fd, uint32_t buf_length, int32_t &ret_val);
+    int32_t doReprocess(int buf_fd, size_t buf_length, int32_t &ret_val);
     int32_t doReprocessOffline(mm_camera_super_buf_t *frame);
     int32_t stop();
 
@@ -141,7 +141,7 @@ private:
     typedef struct {
         QCameraStream *stream;
         cam_mapping_buf_type type;
-        int index;
+        uint32_t index;
     } OfflineBuffer;
 
     uint32_t mSrcStreamHandles[MAX_STREAM_NUM_IN_BUNDLE];
