+++ QCamera2/HAL3/QCamera3PostProc.h
@@ -154,7 +154,7 @@ private:
     uint32_t                   mJpegSessionId;
     uint32_t                   mPostProcMask;
 
-    int8_t                     m_bThumbnailNeeded;
+    uint32_t                   m_bThumbnailNeeded;
     QCamera3Memory             *mJpegMem;
     QCamera3ReprocessChannel *  m_pReprocChannel;
