+++ QCamera2/QCamera2Factory.cpp
@@ -309,7 +309,7 @@ int QCamera2Factory::cameraDeviceOpen(int camera_id,
             delete hw;
         }
     } else if (mHalDescriptors[camera_id].device_version == CAMERA_DEVICE_API_VERSION_1_0) {
-        QCamera2HardwareInterface *hw = new QCamera2HardwareInterface(camera_id);
+        QCamera2HardwareInterface *hw = new QCamera2HardwareInterface((uint32_t)camera_id);
         if (!hw) {
             ALOGE("Allocation of hardware interface failed");
             return NO_MEMORY;
@@ -390,7 +390,7 @@ int QCamera2Factory::openLegacy(
         case CAMERA_DEVICE_API_VERSION_1_0:
         {
             QCamera2HardwareInterface *hw =
-                new QCamera2HardwareInterface(cameraId);
+                new QCamera2HardwareInterface((uint32_t)cameraId);
             if (!hw) {
                 ALOGE("%s: Allocation of hardware interface failed", __func__);
                 return NO_MEMORY;
