+++ QCamera2/stack/mm-camera-test/src/mm_qcamera_reprocess.c
@@ -42,7 +42,7 @@ static void mm_app_reprocess_notify_cb(mm_camera_super_buf_t *bufs,
     int i = 0;
     int rc = 0;
 
-    CDBG_ERROR("%s: BEGIN - length=%d, frame idx = %d\n",
+    CDBG_ERROR("%s: BEGIN - length=%zu, frame idx = %d\n",
          __func__, frame->frame_len, frame->frame_idx);
 
     /* find channel */
@@ -276,7 +276,7 @@ int mm_app_stop_reprocess(mm_camera_test_obj_t *test_obj)
 
 int mm_app_do_reprocess(mm_camera_test_obj_t *test_obj,
                         mm_camera_buf_def_t *frame,
-                        uint8_t meta_idx,
+                        uint32_t meta_idx,
                         mm_camera_super_buf_t *super_buf,
                         mm_camera_stream_t *src_meta)
 {
